pyproject.toml
[project]
name = "btx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.4.3",
    "pyarrow>=21.0.0",
    "ta>=0.11.0",
    "yfinance>=1.0",
]

[dependency-groups]
dev = [
    "isort>=7.0.0",
    "pandas-stubs>=2.3.3.260113",
    "ruff>=0.14.14",
]

[tool.basedpyright]
typeCheckingMode = "strict"
reportMissingTypeStubs = "none"
reportUnknownMemberType = "none"


